requirements.txt
openai-whisper
SpeechRecognition
noisereduce
librosa
soundfile
sounddevice
gtts
nltk
wordcloud
numpy
pandas
matplotlib
scipy
deep-translator
sumy
